+++ ginga/tests/test_iqcalc.py
@@ -2,7 +2,12 @@ import logging
 import numpy as np
 import pytest
 
-from ginga.util import iqcalc
+from ginga.util import iqcalc, iqcalc_astropy
+try:
+    from scipy import optimize  # noqa
+    HAS_SCIPY = True
+except ImportError:
+    HAS_SCIPY = False
 
 try:
     from scipy import optimize  # noqa
@@ -43,31 +48,39 @@ def test_get_median_mask():
 
 @pytest.mark.skipif('not HAS_SCIPY')
 class TestIQCalc:
+    # Shared attributes that subclass can access.
+    logger = logging.getLogger("TestIQCalc")
+    input_arrays = (
+        np.array([0., 0., 11., 12., 8., 9., 37., 96., 289., 786.,
+                  1117., 795., 286., 86., 26., 18., 0., 8., 0., 0.]),
+        np.array([0., 9., 0., 0., 0., 34., 25., 60., 196., 602.,
+                  1117., 1003., 413., 135., 29., 0., 3., 0., 4., 3.]))
 
     def setup_class(self):
-        logger = logging.getLogger("TestIQCalc")
-        self.iqcalc = iqcalc.IQCalc(logger=logger)
+        self.iqcalc = iqcalc.IQCalc(logger=self.logger)
+        self.fwhm_funcs = (self.iqcalc.calc_fwhm_gaussian,
+                           self.iqcalc.calc_fwhm_moffat)
+        self.answers = ((2.8551, 2.7732),  # Gaussian
+                        (2.77949, 2.6735)  # Moffat
+                        )
 
-    def test_fwhm_gaussian(self):
-        """Test FHWM gaussian measuring function in 1D."""
-        x = np.array([0., 0., 11., 12., 8., 9., 37., 96., 289., 786.,
-                      1117., 795., 286., 86., 26., 18., 0., 8., 0., 0.])
-        y = np.array([0., 9., 0., 0., 0., 34., 25., 60., 196., 602.,
-                      1117., 1003., 413., 135., 29., 0., 3., 0., 4., 3.])
+    def test_fwhm(self):
+        """Test FWHM measuring function in 1D."""
+        for i, func in enumerate(self.fwhm_funcs):
+            for j, arr1d in enumerate(self.input_arrays):
+                res = func(arr1d)
+                np.testing.assert_allclose(
+                    res.fwhm, self.answers[i][j], atol=1e-4)
 
-        res = self.iqcalc.calc_fwhm_gaussian(x)
-        assert np.isclose(res.fwhm, 2.8551, atol=1e-04)
-        res = self.iqcalc.calc_fwhm_gaussian(y)
-        assert np.isclose(res.fwhm, 2.7732, atol=1e-04)
 
-    def test_fwhm_moffat(self):
-        """Test FWHM moffat measuring function in 1D."""
-        x = np.array([0., 0., 11., 12., 8., 9., 37., 96., 289., 786.,
-                      1117., 795., 286., 86., 26., 18., 0., 8., 0., 0.])
-        y = np.array([0., 9., 0., 0., 0., 34., 25., 60., 196., 602.,
-                      1117., 1003., 413., 135., 29., 0., 3., 0., 4., 3.])
-
-        res = self.iqcalc.calc_fwhm_moffat(x)
-        assert np.isclose(res.fwhm, 2.77949, atol=1e-04)
-        res = self.iqcalc.calc_fwhm_moffat(y)
-        assert np.isclose(res.fwhm, 2.6735, atol=1e-04)
+class TestIQCalcAstropy(TestIQCalc):
+    def setup_class(self):
+        """Customize for Astropy implementation."""
+        self.iqcalc = iqcalc_astropy.IQCalc(logger=self.logger)
+        self.fwhm_funcs = (self.iqcalc.calc_fwhm_gaussian,
+                           self.iqcalc.calc_fwhm_moffat,
+                           self.iqcalc.calc_fwhm_lorentz)
+        self.answers = ((2.8551, 2.7732),  # Gaussian
+                        (2.77949, 2.6735),  # Moffat
+                        (1.9570, 1.8113)  # Lorentz
+                        )